// src/types/movie.ts
export type Movie = {
  name_french: string
  viewing_date: Date
  grade: number
  comment: string
  name_original: string
  grade_story: number
  grade_directing: number
  grade_acting: number
  grade_music: number
  grade_feeling: number
  grade_personal: number
  watch_place: string[]
  genre: string[]
  tmdb_id: number
  img: string
  release_date: Date
  runtime: number
  cast: string[]
  watch_partner: string[]
  countries: string[]
  director: string[]
  producer: string
  synopsis: string
  tmdb_rating: number
}